Updated the Amount feature from Integer to Boolean

+++ app/src/main/java/com/example/mpos/ui/deals/adaptor/DealsAdaptor.kt
@@ -41,7 +41,7 @@ class DealsAdaptor<T>(private val list: List<T>) :
                     orderList.add(
                         OfferDesc(
                             title = it.description,
-                            price = data.price.toDouble(),
+                            price = data.price,
                             selected = true
                         )
                     )
